+++ Sources/VComponents/Components/Modals/Bottom Sheet/_VBottomSheet.swift
@@ -142,7 +142,7 @@ struct _VBottomSheet<HeaderLabel, Content>: View
 
     @ViewBuilder private var header: some View {
         if hasHeader {
-            HStack(spacing: model.layout.labelCloseButtonSpacing, content: {
+            HStack(alignment: model.layout.headerAlignment, spacing: model.layout.labelCloseButtonSpacing, content: {
                 Group(content: {
                     if model.misc.dismissType.contains(.leadingButton) {
                         closeButton